fix: cambios en nombres, navegacion y rutas
