+++ src/pages/admin/configEmail/accionesPlantillas/acccionesPlantilla.tsx
@@ -1,13 +1,13 @@
 import React from "react"
 import { Button, ButtonGroup, Container, Input, InputGroup, InputGroupText } from "reactstrap"
-import { TypeUtilities } from "../../../../utilities/TypeUtilities"
-import { Fetcher as FetcherAcciones, Selector as SelectorAcciones, Action as ActionAcciones } from "../../../../store/slices/accionesPlantilla"
-import { INIT } from "../../../../store/slices/accionesPlantilla"
-import { useDispatch, useSelector } from "../../../../store"
-import { Tables } from "../../../../components/commons/tables/tables"
+import { TypeUtilities } from "@utilities/TypeUtilities"
+import { Fetcher as FetcherAcciones, Selector as SelectorAcciones, Action as ActionAcciones } from "@store/slices/accionesPlantilla"
+import { INIT } from "@store/slices/accionesPlantilla"
+import { useDispatch, useSelector } from "@store/index"
+import { Tables } from "@components/commons/tables/tables"
 import { faEdit } from "@fortawesome/free-solid-svg-icons"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { ButtonPrimary, ButtonSecondary } from "../../../../components/shared/buttons"
+import { ButtonPrimary, ButtonSecondary } from "@components/shared/buttons"
 import Swal from "sweetalert2"
 
 type AccionesInfo = {
