+++ src/pages/admin/docentePage/crearDocente/creacionDocente.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { InputGroup, InputGroupText, Input } from "reactstrap";
-import { Type as DocentesType } from '../../../../store/slices/docentes/_namespace';
-import { ButtonSecondary } from "../../../../components/shared/buttons";
-import MaskedInput from "../../../../components/shared/inputs";
-import { Fetcher as FetcherFacultades, Selector as SelectorFacultades } from "../../../../store/slices/facultades";
-import { Fetcher as FetcherDocentes, Selector as SelectorDocente } from "../../../../store/slices/docentes";
-import { useDispatch, useSelector } from "../../../../store";
-import { TypeUtilities } from "../../../../utilities/TypeUtilities";
-import { maskDNI, maskPhone } from "../../../../components/shared/inputs/utils/index";
+import { Type as DocentesType } from '@store/slices/docentes/_namespace';
+import { ButtonSecondary } from "@components/shared/buttons";
+import MaskedInput from "@components/shared/inputs";
+import { Fetcher as FetcherFacultades, Selector as SelectorFacultades } from "@store/slices/facultades";
+import { Fetcher as FetcherDocentes, Selector as SelectorDocente } from "@store/slices/docentes";
+import { useDispatch, useSelector } from "@store/index";
+import { TypeUtilities } from "@utilities/TypeUtilities";
+import { maskDNI, maskPhone } from "@components/shared/inputs/utils/index";
 import '../../../admin/crearTernas/_ternas.css'
 import Swal from 'sweetalert2';
