+++ src/pages/admin/pensums/index.tsx
@@ -7,13 +7,13 @@ import {
     TabContent,
     TabPane,
 } from "reactstrap";
-import { useDispatch } from "../../store";
-import { Action as ActionPensum } from '../../store/slices/pensums';
-import { Action as ActionSecciones } from '../../store/slices/secciones';
-import { Action as ActionPeriodo } from '../../store/slices/periodo';
+import { useDispatch } from "@store/index";
+import { Action as ActionPensum } from '@store/slices/pensums';
+import { Action as ActionSecciones } from '@store/slices/secciones';
+import { Action as ActionPeriodo } from '@store/slices/periodo';
 
-import { pensumMenu } from "../../components/shared/nav/subMenus";
-import Breadcrumbs from "../../components/shared/breadcrumb/Breadcrumbs";
+import { pensumMenu } from "@components/shared/nav/subMenus";
+import Breadcrumbs from "@components/shared/breadcrumb/Breadcrumbs";
 
 type TypeBreadcumb = {
     title: string;
