+++ src/pages/admin/pensums/mostrarPensums/coprogramaticas.tsx
@@ -1,23 +1,23 @@
 import React, { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "../../../store";
+import { useDispatch, useSelector } from "@store/index";
 
-import type { Type as TypePensums } from "../../../store/slices/pensums/_namespace";
-import type { Type as TypeSecciones } from "../../../store/slices/secciones/_namespace";
-import type { Type as TypeModals } from "../../../Api/namespaces/modals";
+import type { Type as TypePensums } from "@store/slices/pensums/_namespace";
+import type { Type as TypeSecciones } from "@store/slices/secciones/_namespace";
+import type { Type as TypeModals } from "@root/Api/namespaces/modals";
 
-import { Action as ActionPensum, Fetcher as FetcherPensum, Selector as SelectorPensum } from "../../../store/slices/pensums";
-import { Fetcher as FetcherDocentes, Selector as SelectorDocentes } from '../../../store/slices/docentes';
-import { Action as ActionSecciones, Fetcher as FetcherSecciones, Selector as SelectorSecciones } from '../../../store/slices/secciones';
-import { Fetcher as FetcherPeriodo, Selector as SelectorPeriodos } from "../../../store/slices/periodo";
-import { Fetcher as FetcherFacultad, Selector as SelectorFacultad } from "../../../store/slices/facultades";
+import { Action as ActionPensum, Fetcher as FetcherPensum, Selector as SelectorPensum } from "@store/slices/pensums";
+import { Fetcher as FetcherDocentes, Selector as SelectorDocentes } from '@store/slices/docentes';
+import { Action as ActionSecciones, Fetcher as FetcherSecciones, Selector as SelectorSecciones } from '@store/slices/secciones';
+import { Selector as SelectorFacultad, Fetcher as FetcherFacultad } from '@store/slices/facultades';
+import { Fetcher as FetcherPeriodo, Selector as SelectorPeriodos } from '@store/slices/periodo';
 
 import { Container, Card, Row, Col, Button, CardHeader, ButtonGroup, FormGroup, Label, Input, Form, ModalBody, ModalFooter, Modal, ModalHeader, Spinner, CardBody } from "reactstrap";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faEdit, faPlus, faBan, faCheck } from "@fortawesome/free-solid-svg-icons";
-import NotFound from "../../../components/shared/notFound";
+import NotFound from "@components/shared/notFound";
 import { isEmpty } from "lodash";
 
-import { days, MODALS_TYPES } from "../../../consts"
+import { days, MODALS_TYPES } from "@root/consts"
 
 type ClasesPorBloque = Record<number, TypePensums.ClaseInfo[]>;
 
@@ -61,28 +61,33 @@ const INITIAL_MODAL_STATE: TypeModals.ModalState = {
     type: null
 };
 
-export default function Carreras() {
+export default function Coprogramaticas() {
     const dispatch = useDispatch();
     const [filtroClase, setFiltroClase] = useState<string>('');
     const [clasesFiltradas, setClasesFiltradas] = useState<ClasesPorBloque>({});
     const [clasesPorBloque, setClasesPorBloque] = useState<ClasesPorBloque>({});
-    const [facultadSeleccionada, setFacultadSeleccionada] = useState<string>('IG04001');
-    const [periodoSeleccionado, setPeriodoSeleccionado] = useState<string>('');
     const [modal, setModal] = useState<TypeModals.ModalState>(INITIAL_MODAL_STATE);
     const [formData, setFormData] = useState<ClaseForm>(INITIAL_FORM_STATE);
     const [sectionForm, setSectionForm] = useState<SectionForm>(INITIAL_SECTION_FORM);
+    const [facultadSeleccionada, setFacultadSeleccionada] = useState<string>('IDIN01001');
+    const [periodoSeleccionado, setPeriodoSeleccionado] = useState<string>('');
 
-    const isLoading = useSelector(SelectorPensum.getIsLoading);
-    const isUpdatePensum = useSelector(SelectorPensum.getIsUpdate);
-    const isUpdateSecciones = useSelector(SelectorSecciones.getIsUpdate);
     const clases = useSelector(SelectorPensum.getClases);
     const carreras = useSelector(SelectorPensum.getCarreras);
     const docentes = useSelector(SelectorDocentes.getDocentes);
     const secciones = useSelector(SelectorSecciones.getSecciones);
-    const periodos = useSelector(SelectorPeriodos.getPeriodos);
     const facultades = useSelector(SelectorFacultad.getFacultades);
+    const periodos = useSelector(SelectorPeriodos.getPeriodos);
+    const isLoading = useSelector(SelectorPensum.getIsLoading);
+    const isUpdatePensum = useSelector(SelectorPensum.getIsUpdate);
+    const isUpdateSecciones = useSelector(SelectorSecciones.getIsUpdate);
+
+    const getTipoClase = (facultadId: string) => {
+        return facultadId === 'IDIN01001' ? 2 : 4;
+    };
 
     useEffect(() => {
+        const tipoClase = getTipoClase(facultadSeleccionada);
         if (clases && carreras && periodos && isLoading) {
             dispatch(ActionPensum.setIsLoading(false));
             if (periodos.length > 0)
@@ -90,17 +95,18 @@ export default function Carreras() {
         }
         if (clases === null && carreras === null && periodos === null && !isLoading) {
             dispatch(ActionPensum.setIsLoading(true));
-            dispatch(FetcherPensum.getClases({ url: "/pensum/getPensum?TipoClase=1" }));
+            dispatch(FetcherPensum.getClases({ url: `/pensum/getPensum?TipoClase=${tipoClase}` }));
             dispatch(FetcherPensum.getCarreras({ url: `/pensum/getPensumBy?facultadId=${facultadSeleccionada}` }));
             dispatch(FetcherDocentes.getDocentes({ url: "/docente/getDocentes" }));
             dispatch(FetcherSecciones.getSecciones({ url: `/secciones/getSections?id_periodo=${periodoSeleccionado}` }));
             dispatch(FetcherPeriodo.getPeriodos({ url: "/periodo/get" }));
             dispatch(FetcherFacultad.getFacultades({ url: "/facultad/getFacultades" }))
         }
-    }, [dispatch, clases, carreras, periodos]);
+    }, [dispatch, clases, carreras, periodos, facultadSeleccionada]);
 
     useEffect(() => {
-        dispatch(FetcherPensum.getCarreras({ url: `/pensum/getPensumBy?facultadId=${facultadSeleccionada}` }))
+        dispatch(FetcherPensum.getCarreras({ url: `/pensum/getPensumBy?facultadId=${facultadSeleccionada}` }));
+        dispatch(FetcherPensum.getClases({ url: `/pensum/getPensum?TipoClase=${getTipoClase(facultadSeleccionada)}` }));
     }, [facultadSeleccionada, isUpdatePensum]);
 
     useEffect(() => {
@@ -112,28 +118,24 @@ export default function Carreras() {
 
         const clasesConBloque = clases.map(clase => ({
             ...clase,
-            bloque: carreras.find(c => c.id_clase === clase.id_clase)?.id_bloque
+            bloque: carreras.find(c => c.id_clase === clase.id_clase)?.id_bloque || 1
         }));
 
-        const porBloque = clasesConBloque.reduce<ClasesPorBloque>((acc, clase) => {
-            if (clase.bloque) {
-                if (!acc[clase.bloque]) {
-                    acc[clase.bloque] = [];
-                }
-                acc[clase.bloque].push(clase);
-            }
-            return acc;
-        }, {});
+        const todasLasClases: ClasesPorBloque = {
+            1: clasesConBloque
+        };
 
-        setClasesPorBloque(porBloque);
-        setClasesFiltradas(porBloque);
+        setClasesPorBloque(todasLasClases);
+        setClasesFiltradas(todasLasClases);
     }, [clases, carreras]);
+
 
     const getDayName = (day: number): string => {
         return days[day] || '';
     };
 
     const handleOpenModal = (type: TypeModals.ModalType, bloque?: number, clase?: TypePensums.ClaseInfo) => {
+        const tipoClase = getTipoClase(facultadSeleccionada);
         const newFormData = type === MODALS_TYPES.CREATE
             ? { ...INITIAL_FORM_STATE, id_bloque: bloque || 1, facultadId: facultadSeleccionada }
             : type === MODALS_TYPES.EDIT && clase
@@ -141,7 +143,7 @@ export default function Carreras() {
                     id_clase: clase.id_clase,
                     nombre_clase: clase.nombre_clase,
                     creditos: clase.creditos,
-                    TipoClase: clase.TipoClase,
+                    TipoClase: tipoClase,
                     facultadId: facultadSeleccionada,
                     id_bloque: clase.bloque || 1
                 }
@@ -282,6 +284,7 @@ export default function Carreras() {
         const nuevaFacultad = e.target.value;
         setFacultadSeleccionada(nuevaFacultad);
     };
+
     const handleChangePeriodo = (e: React.ChangeEvent<HTMLInputElement>) => {
         const periodo = e.target.value;
         setPeriodoSeleccionado(periodo);
@@ -354,7 +357,7 @@ export default function Carreras() {
                 </Row>
             </Form>
 
-            {isLoading || isUpdatePensum || isUpdateSecciones ? (
+            {isLoading ? (
                 <div className="text-center my-5">
                     <Spinner color="primary">
                         Loading...
@@ -397,7 +400,7 @@ export default function Carreras() {
                                                 </ButtonGroup>
                                             </CardHeader>
                                             <CardBody>
-                                                {secciones.length > 0 && carreras && secciones
+                                                {secciones
                                                     ?.filter(seccion => {
                                                         const carrera = carreras.find(c => c.id_clase === clase.id_clase);
                                                         return carrera && seccion.id_ccb === carrera.id_ccb;
@@ -521,15 +524,15 @@ export default function Carreras() {
                 </Form>
             </Modal>
 
-            <Modal isOpen={modal.isOpen && [MODALS_TYPES.CREATE_SECCION, MODALS_TYPES.EDIT_SECCION, MODALS_TYPES.DELETE_SECCION].includes(modal.type || '')} toggle={handleCloseModal}>
+            <Modal isOpen={modal.isOpen && [MODALS_TYPES.CREATE_SECCION, MODALS_TYPES.EDIT_SECCION, 'deleteSeccion'].includes(modal.type || '')} toggle={handleCloseModal}>
                 <ModalHeader toggle={handleCloseModal}>
                     {modal.type === MODALS_TYPES.CREATE_SECCION && `Agregar Sección - ${modal.currentClase?.nombre_clase}`}
                     {modal.type === MODALS_TYPES.EDIT_SECCION && `Editar Sección - ${modal.currentClase?.nombre_clase}`}
-                    {modal.type === MODALS_TYPES.DELETE_SECCION && `Eliminar Sección - ${modal.currentClase?.nombre_clase}`}
+                    {modal.type === 'deleteSeccion' && `Eliminar Sección - ${modal.currentClase?.nombre_clase}`}
                 </ModalHeader>
                 <Form onSubmit={handleSectionSubmit}>
                     <ModalBody>
-                        {modal.type === MODALS_TYPES.DELETE_SECCION ? (
+                        {modal.type === 'deleteSeccion' ? (
                             <div className="text-center">
                                 <h5>¿Está seguro que desea eliminar esta sección?</h5>
                                 <p className="mb-0">Sección: {sectionForm.seccion}</p>
@@ -578,9 +581,12 @@ export default function Carreras() {
                                         onChange={handleSectionInputChange}
                                         required
                                     >
-                                        {days && days.map((day, index) => {
-                                            return <option key={day + index + 1} value={index + 1}>{day}</option>
-                                        })}
+                                        <option value="1">Lunes</option>
+                                        <option value="2">Martes</option>
+                                        <option value="3">Miércoles</option>
+                                        <option value="4">Jueves</option>
+                                        <option value="5">Viernes</option>
+                                        <option value="6">Sábado</option>
                                     </Input>
                                 </FormGroup>
                                 <FormGroup>
@@ -593,9 +599,12 @@ export default function Carreras() {
                                         onChange={handleSectionInputChange}
                                         required
                                     >
-                                        {days && days.map((day, index) => {
-                                            return <option key={day + index + 2} value={index + 1}>{day}</option>
-                                        })}
+                                        <option value="1">Lunes</option>
+                                        <option value="2">Martes</option>
+                                        <option value="3">Miércoles</option>
+                                        <option value="4">Jueves</option>
+                                        <option value="5">Viernes</option>
+                                        <option value="6">Sábado</option>
                                     </Input>
                                 </FormGroup>
                                 <FormGroup>
@@ -642,7 +651,7 @@ export default function Carreras() {
                         {modal.type === MODALS_TYPES.EDIT_SECCION && (
                             <Button
                                 color="danger"
-                                onClick={() => setModal(prev => ({ ...prev, type: MODALS_TYPES.DELETE_SECCION as TypeModals.ModalType }))}
+                                onClick={() => setModal(prev => ({ ...prev, type: 'deleteSeccion' }))}
                             >
                                 Eliminar
                             </Button>
@@ -653,7 +662,7 @@ export default function Carreras() {
                         >
                             {modal.type === MODALS_TYPES.CREATE_SECCION && 'Agregar'}
                             {modal.type === MODALS_TYPES.EDIT_SECCION && 'Guardar'}
-                            {modal.type === MODALS_TYPES.DELETE_SECCION && 'Eliminar'}
+                            {modal.type === 'deleteSeccion' && 'Eliminar'}
                         </Button>
                     </ModalFooter>
                 </Form>
