+++ src/pages/home/home.tsx
@@ -10,12 +10,13 @@ import { useSelector } from '@store/index';
 
 export const Home: React.FC = () => {
   const userInfo = useSelector(SelectorUser.getUser);
+  console.log(1 - userInfo.roleId);
   return (
     <>
       <div className="fondo">
         <div className="container mt-5">
           <div className="row text-center">
-            {MenuItems[1 - userInfo.roleId].map((item, index) => (
+            {MenuItems[userInfo.roleId - 1].map((item, index) => (
               <div className="col-md-6 mb-3" key={index}>
                 <Link to={item.pathTo} className="btn btn-dark w-75">
                   {item.title}{" "}
